modified:   frontend/components/Domain/Recipe/RecipePage/RecipePage.vue 	new file:   frontend/components/global/Wakelock.vue 	modified:   frontend/pages/shopping-lists/_id.vue
